requirements.txt
pandas==2.3.3
openpyxl==3.1.5
openai==2.24.0
python-dotenv==1.2.1
